some more tries
now i got the problem, that when the menu gets hidden and shown the sliders lost their properties as well. this is caused by the newly rendered component.
for the map i now set the css width to 100 vw. so it will always be on full device width and with the visible menu its just not seen. i also changed the map center to get it near münster on half view